backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel[asyncio]
asyncpg
httpx
pydantic-settings
redis
pyotp
qrcode
argon2-cffi
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pillow